+++ Source/FakeBot/FakeBot.swift
@@ -288,7 +288,11 @@ class FakeBot: Bot {
             completion(StaticDataProxy(), nil)
         }
     }
-    
+
+    func feed(strategy: FeedStrategy, completion: @escaping PaginatedCompletion) {
+        completion(StaticDataProxy(), nil)
+    }
+
     func feed(identity: Identity, completion: PaginatedCompletion) {
         completion(StaticDataProxy(), nil)
     }