+++ Source/Localization/Text.swift
@@ -70,6 +70,8 @@ enum Text: String, Localizable, CaseIterable {
     case usersInYourNetwork = "Users in your network"
     case goToYourNetwork = "Go to Your Network"
     case emptyHomeFeedMessage = "And it is rather bare! Have you considered following a few users or topics?"
+    case replied = "{{somebody}} replied"
+    case liked = "{{somebody}} liked"
     case startedFollowing = "{{somebody}} started following"
     case stoppedFollowing = "{{somebody}} stopped following"
     case followStats = "Following {{numberOfFollows}} • Followed by {{numberOfFollowers}}"